+++ src/components/FileMenu/SaveAsDialog.js
@@ -18,7 +18,7 @@ const NAME_MAXLENGTH = 230
 export const SaveAsDialog = ({ type, object, onClose, onSaveAs }) => {
     const [name, setName] = useState(
         object?.displayName || object?.name
-            ? i18n.t('{{objectName}} (copy)', {
+            ? i18n.t('{{- objectName}} (copy)', {
                   objectName: object.name,
               })
             : ''
@@ -45,7 +45,6 @@ export const SaveAsDialog = ({ type, object, onClose, onSaveAs }) => {
             <ModalContent>
                 <InputField
                     label={i18n.t('Name')}
-                    required
                     value={name}
                     onChange={({ value }) =>
                         setName(value.substring(0, NAME_MAXLENGTH))